requirements.txt
fastapi
uvicorn
streamlit
requests
paramiko
pytest
pandas
PyYAML
plotly
numpy
scipy